FlashImage: ImageBranding downloads streamlined Due to boxnames being different between image brands this gets complicated. But for now I create an FlashImage reprosatory on our github that contains json files for each model name and as far I could cover it all the boxes we cover are included there as boxname and also as boxname from OpenATV. FlashImage will download the json file and then knows which images are available and where the json file from the brand could be reached.
Aditionally I restored the downloaded images as it was and the backup
images (hopefully I did not miss something here) and ensure when the
images are ordered that downloaded and backup are always at the bottom

Hopefully everyone can help to get the FlashImage reprosatory complete
and hopefully it will not be forked this time.

Now I also directly retreive them from github. Maybe it is also better
to 'clone' them on a download server from the ImageBrands themselves.

I hope this is a kind of final start to 'work-a-round' the boxnames so
every imagebrand is free to use what they want or prefer
Thanks to @Littlesat