SoftwareUpdate Try to resolve spinner and solve fetching the last mod… …ified date
It seems in python 3 you do not need to use getdate anymore. Just get is
sufficiant to retreive the last modified date. But this should not lead
to spinners. Spinners may only occur when the package could not be
derived from the internet. Therefore also add a timeout.
Thanks to @Littlesat